atskirta i kita faila filtro funkcija
